+++ robolectric/src/main/java/org/robolectric/internal/bytecode/ClassInfo.java
@@ -1,13 +1,38 @@
 package org.robolectric.internal.bytecode;
 
 import java.lang.annotation.Annotation;
+import org.objectweb.asm.Opcodes;
+import org.objectweb.asm.tree.AnnotationNode;
+import org.objectweb.asm.tree.ClassNode;
 
-public interface ClassInfo {
-  String getName();
+public class ClassInfo {
+  private final String className;
+  private final ClassNode classNode;
 
-  boolean isInterface();
+  public ClassInfo(String className, ClassNode classNode) {
+    this.className = className;
+    this.classNode = classNode;
+  }
 
-  boolean isAnnotation();
+  public boolean isInterface() {
+    return (classNode.access & Opcodes.ACC_INTERFACE) != 0;
+  }
 
-  boolean hasAnnotation(Class<? extends Annotation> annotationClass);
+  public boolean isAnnotation() {
+    return (classNode.access & Opcodes.ACC_ANNOTATION) != 0;
+  }
+
+  public boolean hasAnnotation(Class<? extends Annotation> annotationClass) {
+    String internalName = "L" + annotationClass.getName().replace('.', '/') + ";";
+    if (classNode.visibleAnnotations == null) return false;
+    for (Object visibleAnnotation : classNode.visibleAnnotations) {
+      AnnotationNode annotationNode = (AnnotationNode) visibleAnnotation;
+      if (annotationNode.desc.equals(internalName)) return true;
+    }
+    return false;
+  }
+
+  public String getName() {
+    return className;
+  }
 }
